Add support to infer column order from file/stream header
Added basic support to infer column order from file/stream header.  The file/stream must have a delimited or fixed-width header row.  The values from the header are matched to the field/properties in the destination POCO.  A possible future enhancement is to leverage an attribute (perhaps the existing FieldCaptionAttribute) to allow fields/properties to have names that differ from the column header, but still allow dynamic resolution of column order.